Rename SharedResourceUtil to SubmissionResourceUtil
