+++ deposit-integration/src/test/java/org/dataconservancy/pass/deposit/messaging/service/SubmissionProcessorIT.java
@@ -21,7 +21,7 @@ import org.dataconservancy.pass.model.Repository;
 import org.dataconservancy.pass.model.RepositoryCopy;
 import org.dataconservancy.pass.model.Submission;
 import org.junit.Test;
-import submissions.SharedResourceUtil;
+import submissions.SubmissionResourceUtil;
 
 import java.io.InputStream;
 import java.net.URI;
@@ -45,7 +45,7 @@ public class SubmissionProcessorIT extends AbstractSubmissionIT {
 
     @Override
     protected InputStream getSubmissionResources() {
-        return SharedResourceUtil.lookupStream(SUBMISSION_RESOURCES);
+        return SubmissionResourceUtil.lookupStream(SUBMISSION_RESOURCES);
     }
 
     @Test
